Make ControllerManager's default args public
